% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rresync-package.R
\docType{package}
\name{rresync-package}
\alias{rresync-package}
\alias{rresync}
\title{rresync - Client for ResourceSync}
\description{
rresync - Client for ResourceSync
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
